Fixed a few minor bugs in sgenoise

+++ include/sge/noise/sample.hpp
@@ -63,19 +63,20 @@ sample(
 						i.get()
 					)
 				) *
-			  _noise.sample(
-				static_cast<result_type>(
-					std::pow(
-						2.0,
-						static_cast<
-							double
-						>(
-							i.get()
+				static_cast<double>(
+				  _noise.sample(
+					static_cast<result_type>(
+						std::pow(
+							2.0,
+							static_cast<
+								double
+							>(
+								i.get()
+							)
 						)
-					)
-					* freq)
-				* _parameters.position().get()
-			);
+						* freq)
+					* _parameters.position().get()
+				));
 	}
 
 	return
